Moved the current tests to a new file
