+++ hail/python/test/hail/methods/test_pca.py
@@ -1,0 +1,167 @@
+import math
+
+import pytest
+import numpy as np
+
+import hail as hl
+from ..helpers import resource, startTestHailContext, stopTestHailContext, fails_local_backend
+
+setUpModule = startTestHailContext
+tearDownModule = stopTestHailContext
+
+
+@fails_local_backend()
+def test_hwe_normalized_pca():
+    mt = hl.balding_nichols_model(3, 100, 50)
+    eigenvalues, scores, loadings = hl.hwe_normalized_pca(mt.GT, k=2, compute_loadings=True)
+
+    assert len(eigenvalues) == 2
+    assert isinstance(scores, hl.Table)
+    scores.count() == 100
+    assert isinstance(loadings, hl.Table)
+
+    _, _, loadings = hl.hwe_normalized_pca(mt.GT, k=2, compute_loadings=False)
+    assert loadings is None
+
+
+@fails_local_backend()
+def test_pca_against_numpy():
+    mt = hl.import_vcf(resource('tiny_m.vcf'))
+    mt = mt.filter_rows(hl.len(mt.alleles) == 2)
+    mt = mt.annotate_rows(AC=hl.agg.sum(mt.GT.n_alt_alleles()),
+                          n_called=hl.agg.count_where(hl.is_defined(mt.GT)))
+    mt = mt.filter_rows((mt.AC > 0) & (mt.AC < 2 * mt.n_called)).persist()
+    n_rows = mt.count_rows()
+
+    def make_expr(mean):
+        return hl.cond(hl.is_defined(mt.GT),
+                       (mt.GT.n_alt_alleles() - mean) / hl.sqrt(mean * (2 - mean) * n_rows / 2),
+                       0)
+
+    eigen, scores, loadings = hl.pca(hl.bind(make_expr, mt.AC / mt.n_called), k=3, compute_loadings=True)
+    hail_scores = scores.explode('scores').scores.collect()
+    hail_loadings = loadings.explode('loadings').loadings.collect()
+
+    assert len(eigen) == 3
+    assert scores.count() == mt.count_cols()
+    assert loadings.count() == n_rows
+
+    assert len(scores.globals) == 0
+    assert len(loadings.globals) == 0
+
+    # compute PCA with numpy
+    def normalize(a):
+        ms = np.mean(a, axis=0, keepdims=True)
+        return np.divide(np.subtract(a, ms), np.sqrt(2.0 * np.multiply(ms / 2.0, 1 - ms / 2.0) * a.shape[1]))
+
+    g = np.pad(np.diag([1.0, 1, 2]), ((0, 1), (0, 0)), mode='constant')
+    g[1, 0] = 1.0 / 3
+    n = normalize(g)
+    U, s, V = np.linalg.svd(n, full_matrices=0)
+    np_scores = U.dot(np.diag(s)).flatten()
+    np_loadings = V.transpose().flatten()
+    np_eigenvalues = np.multiply(s, s).flatten()
+
+    def check(hail_array, np_array):
+        assert len(hail_array) == len(np_array)
+        for i, (left, right) in enumerate(zip(hail_array, np_array)):
+            assert math.isclose(abs(left), abs(right), rel_tol=1e-5), f'mismatch at index {i}: hl={left}, np={right}'
+
+    check(eigen, np_eigenvalues)
+    check(hail_scores, np_scores)
+    check(hail_loadings, np_loadings)
+
+
+@fails_local_backend()
+def test_blanczos_against_numpy():
+
+    def concatToNumpy(field, horizontal=True):
+        blocks = field.collect()
+        if horizontal:
+            return np.concatenate(blocks, axis=0)
+        else:
+            return np.concatenate(blocks, axis=1)
+
+    mt = hl.import_vcf(resource('tiny_m.vcf'))
+    mt = mt.filter_rows(hl.len(mt.alleles) == 2)
+    mt = mt.annotate_rows(AC=hl.agg.sum(mt.GT.n_alt_alleles()),
+                          n_called=hl.agg.count_where(hl.is_defined(mt.GT)))
+    mt = mt.filter_rows((mt.AC > 0) & (mt.AC < 2 * mt.n_called)).persist()
+    n_rows = mt.count_rows()
+
+    def make_expr(mean):
+        return hl.cond(hl.is_defined(mt.GT),
+                       (mt.GT.n_alt_alleles() - mean) / hl.sqrt(mean * (2 - mean) * n_rows / 2),
+                       0)
+
+    k = 3
+
+    float_expr = make_expr(mt.AC / mt.n_called)
+
+    eigens, scores_t, loadings_t = hl._blanczos_pca(float_expr, k=k, q_iterations=7, compute_loadings=True)
+    A = np.array(float_expr.collect()).reshape((3, 4)).T
+    scores = concatToNumpy(scores_t.scores)
+    loadings = concatToNumpy(loadings_t.loadings)
+    scores = np.reshape(scores, (len(scores) // k, k))
+    loadings = np.reshape(loadings, (len(loadings) // k, k))
+
+    assert len(eigens) == 3
+    assert scores_t.count() ==  mt.count_cols()
+    assert loadings_t.count() == n_rows
+    np.testing.assert_almost_equal(A @ loadings, scores)
+
+    assert len(scores_t.globals) == 0
+    assert len(loadings_t.globals) == 0
+
+    # compute PCA with numpy
+    def normalize(a):
+        ms = np.mean(a, axis=0, keepdims=True)
+        return np.divide(np.subtract(a, ms), np.sqrt(2.0 * np.multiply(ms / 2.0, 1 - ms / 2.0) * a.shape[1]))
+
+    g = np.pad(np.diag([1.0, 1, 2]), ((0, 1), (0, 0)), mode='constant')
+    g[1, 0] = 1.0 / 3
+    n = normalize(g)
+    U, s, V = np.linalg.svd(n, full_matrices=0)
+    np_loadings = V.transpose()
+    np_eigenvalues = np.multiply(s, s)
+
+    def bound(vs, us):  # equation 12 from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4827102/pdf/main.pdf
+        return 1/k * sum([np.linalg.norm(us.T @ vs[:,i]) for i in range(k)])
+
+    np.testing.assert_allclose(eigens, np_eigenvalues, rtol=0.05)
+    assert bound(np_loadings, loadings) > 0.9
+
+
+@fails_local_backend()
+def test_blanczos_against_hail():
+    k = 10
+
+    def concatToNumpy(field, horizontal=True):
+        blocks = field.collect()
+        if horizontal:
+            return np.concatenate(blocks, axis=0)
+        else:
+            return np.concatenate(blocks, axis=1)
+
+    hl.utils.get_1kg('data/')
+    hl.import_vcf('data/1kg.vcf.bgz').write('data/1kg.mt', overwrite=True)
+    dataset = hl.read_matrix_table('data/1kg.mt')
+
+    b_eigens, b_scores, b_loadings = hl._blanczos_pca(hl.int(hl.is_defined(dataset.GT)), k=k, q_iterations=3, compute_loadings=True)
+    b_scores = concatToNumpy(b_scores.scores)
+    b_loadings = concatToNumpy(b_loadings.loadings)
+    b_scores = np.reshape(b_scores, (len(b_scores) // k, k))
+    b_loadings = np.reshape(b_loadings, (len(b_loadings) // k, k))
+
+    h_eigens, h_scores, h_loadings = hl.pca(hl.int(hl.is_defined(dataset.GT)), k=k, compute_loadings=True)
+    h_scores = np.reshape(concatToNumpy(h_scores.scores), b_scores.shape)
+    h_loadings = np.reshape(concatToNumpy(h_loadings.loadings), b_loadings.shape)
+
+    # equation 12 from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4827102/pdf/main.pdf
+    def bound(vs, us):
+        return 1/k * sum([np.linalg.norm(us.T @ vs[:,i]) for i in range(k)])
+
+    MEV = bound(h_loadings, b_loadings)
+
+    np.testing.assert_allclose(b_eigens, h_eigens, rtol=0.05)
+    assert MEV > 0.9
